fan working

+++ include/fan.h
@@ -29,7 +29,7 @@ public:
     bool is_enabled() const;
     uint8_t get_speed() const;
     bool is_locked() const; // Check if fan is stalled/locked
-    float get_speed_percentage() const;
+    int get_speed_percentage() const;
 };
 
 class PIDFanController {
